Don't use add_edges_from for layers benchmark
The add*from methods for the graph classes are only recent additions
to retworkx. While they offer better performance for adding multiples to
a graph we shouldn't rely on them for benchmark setup, especially for
methods that predate it. The layers benchmarks were one example, they
were using add_edges_from to add all the edges in one batch, but this
limits the history of that benchmark to when add_edges_from was added to
retworkx instead of when layers() was added. This commit corrects this
by just calling add_edge() in a loop.