Improve compatibility of renamed function benchmarks
There were several functions that were originally merged as
dag_*() when PyDAG was the only class in retworkx. They were renamed as
part of the rename of PyDAG to PyDiGraph in Qiskit/retworkx#88 and were
never actually released as the dag_*() name.
However to get the historical data for the function from commits prior
to the rename we would need to call the old name. This commit updates
the benchmark to use dag_* names if they're present, otherwise we'll
have no data from commits prior to the rename in Qiskit/retworkx#88
which might have useful insights.