Add node benchmarks for NYC road map graph
This commit adds an example graph sourced from the 9th DIMACS
challenge [1] (thanks to @t-imamichi for the suggestion), specifically
the smallest real world example provided [2] which is a graph of the
New York City road map with edge weight as physical distance (as opposed
to travel time). [3] This serves as a good benchmark case because it
represents a large example graph to benchmark operations on and also
represents a real world problem instead of just a randomly generated
graph of arbitrary size. A parser function is then added to read the
input .gr file and return a retworkx graph object.

A first set of benchmarks are added for timing simple node operations
on this example graph (get the number of nodes, node data, etc). This is
just a starting point in future commits we'll both expand the number of
benchmarks using this example and also add additional example graphs.

[1] http://users.diag.uniroma1.it/challenge9/
[2] http://users.diag.uniroma1.it/challenge9/download.shtml
[3] http://users.diag.uniroma1.it/challenge9/data/USA-road-d/USA-road-d.NY.gr.gz