Add benchmarks for paths and preds/succ with NYC road map
This commit adds benchmarks that time the path, predecessor, and
successor functions using the NYC road map from the 9th DIMACS challenge
which was added in commit 74d7c49055246330918cfb8b8e1edc4a9a05aa6d. This
provides a more realistic example for the runtime of these functions
with a real world graph in addition to a synthetically constructed one.