Split up node creation and add remove_nodes_from bench
The node creation benchmarks were attempting to run benchmarks with both
PyDiGraph/PyDAG and PyGraph in the same class. This however prevented
getting historical timing data from before the introduction of PyGraph
since the setup will fail even if we weren't using PyGraph for a
particular run. This commit addresses this by splitting out the

At the same time a new benchmark method is added for the recently added
remove_nodes_from() method.