Add bechmarks for Asia road map graph
This commit adds an example graph sourced from the 10th DIMACS
challenge [1] graph files [2]. Specifically the Asia road map example
graph file [3] which is in the METIS 4.0 graph file format. This serves
as an even larger real world problem (compared to the NYC road map
already in the benchmarks) with 11,950,757 nodes and 12,711,603 edges.

[1] https://www.cc.gatech.edu/dimacs10/index.shtml
[2] https://www.cc.gatech.edu/dimacs10/downloads.shtml
[3] https://www.cc.gatech.edu/dimacs10/archive/streets.shtml