Add benchmarks with Western USA Roadmap graph
This commit adds new benchmarks based on the Western USA Roadmap
benchmark graph from the 9th DIMACS challenge. [1] While the recent
large graph addtions were only of an undirected graph this adds a larger
example with a directed case too. Additionally, this marks the first use
of git-lfs. This is an experiment to see if this will enable the use of
much larger graph files for benchmarking. Right now we're stuck at the
100MB file size limit that github puts on it's repos hopefully this
should enable even larger graph files to be used for benhcmarking.

[1] http://users.diag.uniroma1.it/challenge9/download.shtml