Rewrite most benchmarks
The existing benchmarks had a lot of issues including that many weren't
testing good graphs. This is an attempt to clean a bunch of that up and
improves many of the benchmarks. This will invalidate most results and
those benchmarks will need to be rerun. Because of this the python
version is bumped to the latest release to enable running with python
3.9